Update from main

+++ Assets/Scripts/MapGenerator/Tile.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class Tile
 {
-    public (int x, int y) position; // position info
-    public int type; // 0 initialized, 1 path, 2 start, 3 end
+    public (byte x, byte y) position; // position info
+    public byte type; // 0 initialized, 1 path, 2 start, 3 end
 
     public Tile() { }
-    public Tile((int x, int y) newPosition, int newType)
+    public Tile((byte x, byte y) newPosition, byte newType)
     {
         position = newPosition;
         type = newType;
